moved _validate_vartypes as well
